YP_FINAL_V2/requirements.env.txt
fastapi
uvicorn[standard]
sqlalchemy
lyricsgenius
numpy
torch
transformers
sentence-transformers
requests
backoff
scipy
PyQt5
nltk
sentencepiece
cryptography
python-dotenv
huggingface_hub[hf_xet]
faiss-cpu
np






